Update upstream dependencies
Notably, set a minimum version requirement for appstream-util. Prior to
0.7.2, `appstream-util validate` would complain if your application ID
did not end in `.desktop`. That requirement was dropped as a result of
https://github.com/hughsie/appstream-glib/issues/146;
transmission-remote-gtk began taking advantage of the looser checking
in dd01b099b7f1807999a08b09bbe06ed2d65aeadd.

Also add an explicit dependency on libappstream-glib-dev, which was previously
being pulled in as a transitive dependency.